+++ include/PolygonFitting/PolygonFitting.h
@@ -30,32 +30,32 @@ public:
 
   void checkIfValidFit();
 
-  bool findBestScale( double &maxScale, Vertex &bestFitOrigin );
+  bool findBestScale( float &maxScale, Vertex &bestFitOrigin );
 
   bool isValidFit( Vertex &c );
 
-  void drawImageFit( double maxScale, Vertex bestFitOrigin );
+  void drawImageFit( float maxScale, Vertex bestFitOrigin );
 };
 
-void findBestFit( Polygon P2, Polygon R, double &scale, double &rotation, Vertex &offset, double rotationOffset = 1.0, double startRotation = 0.0 );
+void findBestFit( Polygon P2, Polygon R, float &scale, float &rotation, Vertex &offset, float rotationOffset = 1.0, float startRotation = 0.0 );
 
-void drawImage( Polygon P, Polygon R, double scale, double rotation, Vertex offset, string outputName, double scaleUp );
+void drawImage( Polygon P, Polygon R, float scale, float rotation, Vertex offset, string outputName, float scaleUp );
 
-void drawImage( string imageName, Polygon R, double scale, double rotation, Vertex offset, string outputName, double scaleUp );
+void drawImage( string imageName, Polygon R, float scale, float rotation, Vertex offset, string outputName, float scaleUp );
 
-void removeExtrasP( Polygon &P, double epsilon = -1 );
+void removeExtrasP( Polygon &P, float epsilon = -1 );
 
-void polygonFromAlphaImage( Polygon &P, string imageName, double resize = 1.0 );
+void polygonFromAlphaImage( Polygon &P, string imageName, float resize = 1.0 );
 
-void concavePolygonFromAlphaImage( Polygon &P, string imageName, double resize );
+void concavePolygonFromAlphaImage( Polygon &P, string imageName, float resize );
 
 void convexPolygonFromVertices( vector< Vertex > &vertices, Polygon &P );
 
-double ccw( Vertex &v1, Vertex &v2, Vertex &v3 );
+float ccw( Vertex &v1, Vertex &v2, Vertex &v3 );
 
 bool minY(Vertex i, Vertex j);
 
-double distance( Vertex v1, Vertex v2 );
+float distance( Vertex v1, Vertex v2 );
 
 bool grahamScanSort ( Vertex v1, Vertex v2, Vertex p );
 
@@ -63,12 +63,12 @@ vector< Edge > grahamScan( vector< Vertex > vertices );
 
 Edge reverseEdge( Edge e );
 
-pair< double, double > intersection( Vertex &v1, Vertex &v2, Vertex &v3, Vertex &v4 );
+pair< float, float > intersection( Vertex &v1, Vertex &v2, Vertex &v3, Vertex &v4 );
 
-Vertex intersectionPoint( double t, Vertex &v1, Vertex &v2 );
+Vertex intersectionPoint( float t, Vertex &v1, Vertex &v2 );
 
 bool isOnLineSegment( Vertex v1, Vertex v2, Vertex v3 );
 
-double intersectionEdges( Vertex v1, Vertex v2, Vertex v3, Vertex v4, Vertex v5, Vertex v6, Vertex o1, Vertex o2, Vertex o3 );
+float intersectionEdges( Vertex v1, Vertex v2, Vertex v3, Vertex v4, Vertex v5, Vertex v6, Vertex o1, Vertex o2, Vertex o3 );
 
-void decimate( Polygon &P, double epsilon );
+void decimate( Polygon &P, float epsilon );
