+++ lib/PolygonFitting/Polygon.c++
@@ -20,7 +20,7 @@ Polygon& Polygon::operator=( const Polygon& rhs )
   return *this;
 }
 
-void Polygon::addVertex( double x, double y )
+void Polygon::addVertex( float x, float y )
 {
   vertices.push_back( Vertex(x,y) );
 }
@@ -52,7 +52,7 @@ void Polygon::switchDirection()
 
 void Polygon::makeClockwise()
 {
-  double n = 0;
+  float n = 0;
   for( int i = 0; i < edges.size(); ++i )
   {
     n += ( vertices[edges[i].v2].x - vertices[edges[i].v1].x ) * ( vertices[edges[i].v2].y + vertices[edges[i].v1].y );
@@ -64,7 +64,7 @@ void Polygon::makeClockwise()
   }
 }
 
-void Polygon::scaleBy( double scale )
+void Polygon::scaleBy( float scale )
 {
   for( int i = 0; i < vertices.size(); ++i )
   {
@@ -81,9 +81,9 @@ void Polygon::offsetBy( Vertex offset )
   }
 }
 
-void Polygon::rotateBy( double r, bool offset )
+void Polygon::rotateBy( float r, bool offset )
 {
-  double pi = 3.14159265358979;
+  float pi = 3.14159265358979;
 
   r *= pi / 180.0;
 
@@ -91,8 +91,8 @@ void Polygon::rotateBy( double r, bool offset )
 
   for( int i = 0; i < vertices.size(); ++i )
   {
-    double x = vertices[i].x * cos( r ) - vertices[i].y * sin( r );
-    double y = vertices[i].x * sin( r ) + vertices[i].y * cos( r );
+    float x = vertices[i].x * cos( r ) - vertices[i].y * sin( r );
+    float y = vertices[i].x * sin( r ) + vertices[i].y * cos( r );
     vertices[i].x = x;
     vertices[i].y = y;
 
@@ -108,7 +108,7 @@ void Polygon::rotateBy( double r, bool offset )
 
 Vertex Polygon::center()
 {
-  double maxX = -1000000000, minX = 1000000000, maxY = -1000000000, minY = 1000000000;
+  float maxX = -1000000000, minX = 1000000000, maxY = -1000000000, minY = 1000000000;
 
   for( int i = 0; i < vertices.size(); ++i )
   {
@@ -127,21 +127,21 @@ Vertex Polygon::center()
   return Vertex((minX+maxX)/2,(minY+maxY)/2);
 }
 
-Vertex Polygon::computeOffset( double r )
+Vertex Polygon::computeOffset( float r )
 {
-  double pi = 3.14159265358979;
+  float pi = 3.14159265358979;
 
   r *= pi / 180.0;
 
-  double px = -1000000, py = -1000000, bx = -1000000, by = -1000000;
+  float px = -1000000, py = -1000000, bx = -1000000, by = -1000000;
 
-  double xRange[2] = {1000000000,-1000000000};
-  double yRange[2] = {1000000000,-1000000000};
+  float xRange[2] = {1000000000,-1000000000};
+  float yRange[2] = {1000000000,-1000000000};
 
   for( int i = 0; i < vertices.size(); ++i )
   {
-    double x = vertices[i].x * cos( r ) - vertices[i].y * sin( r );
-    double y = vertices[i].x * sin( r ) + vertices[i].y * cos( r );
+    float x = vertices[i].x * cos( r ) - vertices[i].y * sin( r );
+    float y = vertices[i].x * sin( r ) + vertices[i].y * cos( r );
 
     xRange[0] = min(vertices[i].x,xRange[0]);
     xRange[1] = max(vertices[i].x,xRange[1]);
@@ -163,8 +163,8 @@ Vertex Polygon::computeOffset( double r )
   { 
     for( int j = 0; j < 2; ++j )
     {
-      double x = xRange[j] * cos( r ) - yRange[i] * sin( r );
-      double y = xRange[j] * sin( r ) + yRange[i] * cos( r );
+      float x = xRange[j] * cos( r ) - yRange[i] * sin( r );
+      float y = xRange[j] * sin( r ) + yRange[i] * cos( r );
 
       bx = max( -x, bx );
       by = max( -y, by );
@@ -174,24 +174,24 @@ Vertex Polygon::computeOffset( double r )
   return Vertex( px - bx, py - by );
 }
 
-double Polygon::minRatio()
+float Polygon::minRatio()
 {
-  double largestRatio = -1000000000;
-//  double bestAngle = 0;
+  float largestRatio = -1000000000;
+//  float bestAngle = 0;
 
-//  double lw, lh;
+//  float lw, lh;
 
   for( int i = 0; i < edges.size(); ++i )
   {
-    double r = -atan2(vertices[edges[i].v2].y-vertices[edges[i].v1].y,vertices[edges[i].v2].x-vertices[edges[i].v1].x);
+    float r = -atan2(vertices[edges[i].v2].y-vertices[edges[i].v1].y,vertices[edges[i].v2].x-vertices[edges[i].v1].x);
 
-    double xRange[2] = {1000000000,-1000000000};
-    double yRange[2] = {1000000000,-1000000000};
+    float xRange[2] = {1000000000,-1000000000};
+    float yRange[2] = {1000000000,-1000000000};
 
     for( int j = 0; j < vertices.size(); ++j )
     {
-      double x = vertices[j].x * cos( r ) - vertices[j].y * sin( r );
-      double y = vertices[j].x * sin( r ) + vertices[j].y * cos( r );
+      float x = vertices[j].x * cos( r ) - vertices[j].y * sin( r );
+      float y = vertices[j].x * sin( r ) + vertices[j].y * cos( r );
 
       xRange[0] = min(x,xRange[0]);
       xRange[1] = max(x,xRange[1]);
@@ -200,10 +200,10 @@ double Polygon::minRatio()
       yRange[1] = max(y,yRange[1]);
     }
 
-    double width = xRange[1] - xRange[0];
-    double height = yRange[1] - yRange[0];
+    float width = xRange[1] - xRange[0];
+    float height = yRange[1] - yRange[0];
 
-    double ratio = max(width,height)/min(width,height);
+    float ratio = max(width,height)/min(width,height);
 
     largestRatio = max(ratio,largestRatio);
   }
@@ -211,21 +211,21 @@ double Polygon::minRatio()
   return largestRatio;
 }
 
-double Polygon::minLength()
+float Polygon::minLength()
 {
-  double minLength = 1000000000;
+  float minLength = 1000000000;
 
   for( int i = 0; i < edges.size(); ++i )
   {
-    double r = -atan2(vertices[edges[i].v2].y-vertices[edges[i].v1].y,vertices[edges[i].v2].x-vertices[edges[i].v1].x);
+    float r = -atan2(vertices[edges[i].v2].y-vertices[edges[i].v1].y,vertices[edges[i].v2].x-vertices[edges[i].v1].x);
 
-    double xRange[2] = {1000000000,-1000000000};
-    double yRange[2] = {1000000000,-1000000000};
+    float xRange[2] = {1000000000,-1000000000};
+    float yRange[2] = {1000000000,-1000000000};
 
     for( int j = 0; j < vertices.size(); ++j )
     {
-      double x = vertices[j].x * cos( r ) - vertices[j].y * sin( r );
-      double y = vertices[j].x * sin( r ) + vertices[j].y * cos( r );
+      float x = vertices[j].x * cos( r ) - vertices[j].y * sin( r );
+      float y = vertices[j].x * sin( r ) + vertices[j].y * cos( r );
 
       xRange[0] = min(x,xRange[0]);
       xRange[1] = max(x,xRange[1]);
@@ -234,8 +234,8 @@ double Polygon::minLength()
       yRange[1] = max(y,yRange[1]);
     }
 
-    double width = xRange[1] - xRange[0];
-    double height = yRange[1] - yRange[0];
+    float width = xRange[1] - xRange[0];
+    float height = yRange[1] - yRange[0];
 
     minLength = min(minLength,min(width,height));
   }
@@ -243,7 +243,7 @@ double Polygon::minLength()
   return minLength;
 }
 
-bool Polygon::notOnEdge( double width, double height )
+bool Polygon::notOnEdge( float width, float height )
 {
   for( int j = 0; j < vertices.size(); ++j )
   {
@@ -260,17 +260,17 @@ bool Polygon::notOnEdge( double width, double height )
   return true;
 }
 
-double Polygon::area()
+float Polygon::area()
 {
-  double area = 0;
+  float area = 0;
 
   for( int i=0; i < vertices.size(); i += 2 )
   {
-    double x0 = vertices[i].x;
-    double y0 = vertices[i].y;
-    double x1 = vertices[(i+2) % vertices.size()].x;
-    double y1 = vertices[(i+2) % vertices.size()].y;
-    double a = x0*y1 - x1*y0;
+    float x0 = vertices[i].x;
+    float y0 = vertices[i].y;
+    float x1 = vertices[(i+2) % vertices.size()].x;
+    float y1 = vertices[(i+2) % vertices.size()].y;
+    float a = x0*y1 - x1*y0;
     area += a;
   }
