+++ lib/PolygonFitting/Vertex.c++
@@ -1,6 +1,6 @@
 #include "PolygonFitting/Vertex.h"
 
-Vertex::Vertex( double _x, double _y )
+Vertex::Vertex( float _x, float _y )
 {
   x = _x;
   y = _y;
@@ -20,7 +20,7 @@ Vertex& Vertex::operator=( const Vertex& rhs )
   return *this;
 }
 
-bool Vertex::equals( const Vertex& rhs, double epsilon )
+bool Vertex::equals( const Vertex& rhs, float epsilon )
 {
   return ( abs( this->x - rhs.x ) < epsilon && abs( this->y - rhs.y ) < epsilon );
 }
@@ -35,7 +35,7 @@ bool Vertex::operator!=( const Vertex& rhs )
   return ( abs( this->x - rhs.x ) > 0.0000000001 || abs( this->y - rhs.y ) > 0.0000000001 );
 }
 
-Vertex Vertex::offset( double _x, double _y )
+Vertex Vertex::offset( float _x, float _y )
 {
   return Vertex( x + _x, y + _y );
 }
@@ -45,7 +45,7 @@ Vertex Vertex::offset( Vertex v )
   return Vertex( x + v.x, y + v.y );
 }
 
-Vertex Vertex::scale( double s, Vertex o )
+Vertex Vertex::scale( float s, Vertex o )
 {
   Vertex v(this->x,this->y);
   v = v.offset(Vertex(-o.x,-o.y));
